refactor: moved datadirs to files module
