+++ src/commands/install.rs
@@ -4,7 +4,7 @@ use std::path::{Path, PathBuf};
 
 use crate::{
     commands::{self, download::download_binary},
-    datadirs,
+    files::datadirs,
     files::{archives, filesys},
     github::client::{get_asset, get_release},
     selector::is_env_compatible,
