+++ src/files/mod.rs
@@ -1,3 +1,4 @@
 pub mod archives;
+pub mod datadirs;
 pub mod filesys;
 pub mod magic;
